+ created ssr and ssg pages
