+++ homeworks/andrew-kondratiyk/21-next-js-1/src/pages/csr.tsx
@@ -1,0 +1,7 @@
+import HomePage from '@/components/home-page/HomePage';
+
+const CSR = () => {
+  return <HomePage initialAvatars={[]} />;
+};
+
+export default CSR;
