+++ homeworks/andrew-kondratiyk/21-next-js-1/src/pages/index.tsx
@@ -1,58 +1,23 @@
-import {getAvatars} from '@/api';
-import Avatar from '@/components/avatar/Avatar';
-import styles from '@/styles/Home.module.css';
-import {type AvatarResponse} from '@/types';
-import {useState} from 'react';
+import Link from 'next/link';
 
-const Home = () => {
-  const [avatars, setAvatars] = useState<AvatarResponse[]>([]);
-  const addAvatar = async () => {
-    const avatar = (await getAvatars())[0];
-    setAvatars([...avatars, avatar]);
-  };
-
-  const refreshAvatar = async (custom_id: string) => {
-    const newAvatar = (await getAvatars())[0];
-    const updatedAvatars = avatars.map(avatar =>
-      avatar.custom_id === custom_id ? newAvatar : avatar,
-    );
-    setAvatars(updatedAvatars);
-  };
-
-  const refreshAllAvatars = async () => {
-    const avatarsCount = avatars.length;
-    if (!avatarsCount) {
-      return;
-    }
-    const newAvatars = await getAvatars(avatarsCount);
-    setAvatars(newAvatars);
-  };
-
-  const deleteAvatar = (custom_id: string) => {
-    const updatedAvatars = avatars.filter(
-      avatar => avatar.custom_id !== custom_id,
-    );
-    setAvatars(updatedAvatars);
-  };
-
+export default function Home() {
   return (
-    <main className={styles.main}>
-      <div className={styles.container}>
-        {avatars.map(avatar => (
-          <Avatar
-            key={avatar.custom_id}
-            avatar={avatar}
-            refreshAvatar={() => refreshAvatar(avatar.custom_id)}
-            deleteAvatar={() => deleteAvatar(avatar.custom_id)}
-          />
-        ))}
-        <button className={styles.addAvatarBtn} onClick={addAvatar}></button>
-      </div>
-      <button className={styles.refreshAll} onClick={refreshAllAvatars}>
-        Refresh all
-      </button>
-    </main>
-  );
-};
+    <>
+      <main>
+        <div>
+          <div style={{marginBottom: '20px'}}>
+            <Link href="/ssg">SSG</Link>
+          </div>
+          <div style={{marginBottom: '20px'}}>
+            <Link href="/ssr">SSR</Link>
+          </div>
+          <div>
+            <Link href="/csr">CSR</Link>
+          </div>
+        </div>
 
-export default Home;
+        <h1>Welcome To The Home Page</h1>
+      </main>
+    </>
+  );
+}
